refactor: move detection primitives from src/proxy to src/enforcement
Detectors, injection patterns, sensitive-read, and the content half of
response-inspector become the shared enforcement core. extractResponseText
(JSON-RPC-shape-specific) relocates to proxy/jsonrpc.ts. No behavior change:
full suite passes at the 2393-test baseline.
