+++ src/enforcement/detectors/structured.ts
