+++ src/enforcement/injection-patterns.ts
