+++ src/enforcement/response-inspector.ts
@@ -107,42 +107,6 @@ function redactRanges(text: string, ranges: Array<[number, number]>): string {
   return out;
 }
 
-/**
- * Extract the concatenated text content of an MCP `tools/call` result.
- *
- * Handles the standard shape `{ content: [{ type: 'text', text: '...' }] }`
- * (non-text content items — images, resources — are ignored) as well as a
- * bare string result. Any other/garbage shape yields `''`; this function
- * never throws.
- */
-export function extractResponseText(result: unknown): string {
-  try {
-    if (typeof result === 'string') return result;
-
-    if (result && typeof result === 'object' && !Array.isArray(result)) {
-      const content = (result as { content?: unknown }).content;
-      if (Array.isArray(content)) {
-        let out = '';
-        for (const item of content) {
-          if (
-            item &&
-            typeof item === 'object' &&
-            (item as { type?: unknown }).type === 'text' &&
-            typeof (item as { text?: unknown }).text === 'string'
-          ) {
-            out += (item as { text: string }).text;
-          }
-        }
-        return out;
-      }
-    }
-
-    return '';
-  } catch {
-    return '';
-  }
-}
-
 /** Extract candidate hostnames from `https?://host` URLs and bare `host.tld` tokens. */
 export function extractHosts(text: string): string[] {
   const hosts = new Set<string>();
