+++ tests/unit/proxy-policy.test.ts
@@ -4,7 +4,7 @@ import * as path from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { evaluateCall, evaluateResponse, loadPolicy } from '../../src/proxy/policy.js';
-import type { InspectionResult } from '../../src/proxy/response-inspector.js';
+import type { InspectionResult } from '../../src/enforcement/response-inspector.js';
 
 let tmpDir: string;
