+++ tests/unit/proxy-provenance.test.ts
@@ -4,7 +4,7 @@ import { SessionProvenance, DEFAULT_MAX_TAINT_ENTRIES } from '../../src/proxy/pr
 import type { TaintTag } from '../../src/proxy/provenance.js';
 import { evaluateCall, evaluateResponse } from '../../src/proxy/policy.js';
 import type { EvalContext, ProxyPolicy } from '../../src/proxy/policy.js';
-import type { InspectionResult, ResponseFinding } from '../../src/proxy/response-inspector.js';
+import type { InspectionResult, ResponseFinding } from '../../src/enforcement/response-inspector.js';
 
 // ─────────────────────────────────────────────────────────────────────────
 // Test helpers
